Added try catch block to see the error

+++ app/Http/Controllers/api/OrderController.php
@@ -8,6 +8,7 @@ use Illuminate\Http\Request;
 use App\Models\Order;
 use App\Http\Resources\OrderResource;
 use App\Http\Resources\OrderCollection;
+use Illuminate\Http\Response;
 
 class OrderController extends Controller
 {
@@ -18,7 +19,16 @@ class OrderController extends Controller
      */
     public function index()
     {
-        return new OrderCollection(Order::all());
+        try {
+            return new OrderCollection(Order::all());
+        } catch (\Exception $e) {
+            return response()->json(
+                [
+                    'message' => $e->getMessage()
+                ],
+                Response::HTTP_INTERNAL_SERVER_ERROR
+            );
+        }
     }
 
     /**
@@ -29,18 +39,27 @@ class OrderController extends Controller
      */
     public function store(OrderRequest $request)
     {
-        $order = Order::create([
-            'customer_id' => $request->customer_id,
-            'paid' => $request->paid
-        ]);
+        try {
+            $order = Order::create([
+                'customer_id' => $request->customer_id,
+                'paid' => $request->paid
+            ]);
 
-        if ($order) {
-            $data['data'] = [
-                'id' => $order->id,
-                'message' => 'success'
-            ];
+            if ($order) {
+                $data['data'] = [
+                    'id' => $order->id,
+                    'message' => 'success'
+                ];
 
-            return response()->json($data);
+                return response()->json($data);
+            }
+        } catch (\Exception $e) {
+            return response()->json(
+                [
+                    'message' => $e->getMessage()
+                ],
+                Response::HTTP_INTERNAL_SERVER_ERROR
+            );
         }
     }
 
@@ -52,7 +71,14 @@ class OrderController extends Controller
      */
     public function show($id)
     {
-        return new OrderResource(Order::findOrFail($id));
+        $order = Order::find($id);
+        if ($order) {
+            return new OrderResource($order);
+        } else {
+            return response()->json([
+                'message' => 'Order does not exist'
+            ], Response::HTTP_NOT_FOUND);
+        }
     }
 
     /**
@@ -64,19 +90,24 @@ class OrderController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $order = Order::findOrFail($id);
+        $order = Order::find($id);
+        if ($order) {
+            $order->customer_id = $request->customer_id;
+            $order->paid = $request->paid;
 
-        $order->customer_id = $request->customer_id;
-        $order->paid = $request->paid;
+            $order->save();
 
-        $order->save();
+            $data = [
+                'data' => $order,
+                'message' => 'updated successfully'
+            ];
 
-        $data = [
-            'data' => $order,
-            'message' => 'updated successfully'
-        ];
-
-        return response()->json($data);
+            return response()->json($data);
+        } else {
+            return response()->json([
+                'message' => 'Order does not exist'
+            ], Response::HTTP_NOT_FOUND);
+        }
     }
 
     /**
@@ -87,14 +118,20 @@ class OrderController extends Controller
      */
     public function destroy($id)
     {
-        $order = Order::findOrFail($id);
-        $order->delete();
+        $order = Order::find($id);
+        if ($order) {
+            $order->delete();
 
-        $data['data'] = [
-            'id' => $id,
-            'message' => 'deleted successfully'
-        ];
+            $data['data'] = [
+                'id' => $id,
+                'message' => 'deleted successfully'
+            ];
 
-        return response()->json($data);
+            return response()->json($data);
+        } else {
+            return response()->json([
+                'message' => 'Order does not exist'
+            ], Response::HTTP_NOT_FOUND);
+        }
     }
 }
